tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-originA-B-A.html

<!DOCTYPE html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src=/common/get-host-info.sub.js></script>

</head>

<body>
  <script>
    const loadChildFrame = () => {
      return new Promise(resolve => {

        const crossOriginChildFrame = document.createElement('iframe');

        // Wait for child frame to report number of performance entries.
        window.addEventListener("message", (e) => {
          resolve(e.data);
        });

        crossOriginChildFrame.src = get_host_info().REMOTE_ORIGIN +
          '/performance-timeline/resources/include-frames-subframe.html?origin=' + get_host_info().ORIGIN;

        document.body.appendChild(crossOriginChildFrame);
      });
    }

    promise_test(async () => {

      performance.clearResourceTimings();

      // Load a child frame. The child frame upon loading would load a child frame of its own.
      childFrameEntrySize = await loadChildFrame();

      // Verify the number of performance entries in the child frame.
      assert_equals(childFrameEntrySize, 4, 'Child frame entries should be 4.');

      const entries = performance.getEntries(true);

      const navigationEntries = performance.getEntriesByType(
        'navigation', true);

      const markedEntries = performance.getEntriesByName('entry-name', undefined, true);

      // 3 entries for parent, 0 for child, 2 for grandchild.
      assert_equals(entries.length, 5, 'Total entries should be 5.');

      // 1 entry for parent, 1 for grandchild.
      assert_equals(navigationEntries.length, 2, 'Navigation entries should be 2.');

      // 1 entry for grandchild.
      assert_equals(markedEntries.length, 1, 'Mark entries should be 1.');
    }, 'GetEntries of a document of origin A, its child frame of origin B and \
    its grandchild frame of origin A.');
  </script>
</body>
