third_party/blink/web_tests/platform/mac-mac11/external/wpt/performance-timeline/tentative/include-frames-originA-B-A-expected.txt

This is a testharness.js-based test.
FAIL GetEntries of a document of origin A, its child frame of origin B and     its grandchild frame of origin A. assert_equals: Total entries should be 5. expected 5 but got 6
Harness: the test ran to completion.
